+++ PKHeX.Drawing.PokeSprite/Util/SpriteUtil.cs
@@ -215,8 +215,8 @@ public static class SpriteUtil
     public static int GetDisplayGender(IEncounterTemplate enc) => enc switch
     {
         EncounterSlotGO g => (int)g.Gender & 1,
-        EncounterStatic s => Math.Max(0, s.Gender),
-        EncounterTrade t => Math.Max(0, t.Gender),
+        EncounterStatic s => Math.Max(0, (int)s.Gender),
+        EncounterTrade t => Math.Max(0, (int)t.Gender),
         _ => 0,
     };
